+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
-- Added new ways of wallet generation - from public/private key, from entropy, from secret
+### Added:
+- Add new ways of wallet generation - from public/private key, from entropy, from secret
+- Changed wallet constructor arguments to optional for easier inferencing
 
 ## [1.7.0] - 2022-10-12
 ### Added: